Extract common Template class into Winrm::Template.
